Another nice patch from Mark Cave-Ayland:
This patch corrects a couple of (minor) bugs in the Forth source 
debugger; firstly the number of wordlists is held in #order, rather than 
#vocs which is the total number of wordlists which can be held 
simultaneously.

Secondly, the initial interruptforth setting in kernel/bootstrap.c is 
actually wrong, but since this is reset by enterforth() it is not needed 
anyway. Hence this line can simply be removed.



git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@623 f158a5a8-5612-0410-a976-696ce0be7e32