ppc: Fix /memory reg property
Physical address and size were both using two cells instead of one,
and their order was wrong. This would happen to work for RAM < 4 GB,
since the clear high address bits would happen to match the zero-based
memory location. It would indicate a second bank of size zero though.

Switch the value order and use encode-phys in preparation for ppc64.

This tidies the output of the Haiku boot loader.

v2:
* Document TODOs for ppc64, requested by Alex.

Cc: Alexander Graf <agraf@suse.de>
Signed-off-by: Andreas Färber <andreas.faerber@web.de>


git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@947 f158a5a8-5612-0410-a976-696ce0be7e32