forth: Fix ([IF]) to handle nested [IFDEF] as well
Depending on the condition, either the [IFDEF] FOO or the [ELSE]
would get compiled as an [IF], eating words until [ELSE] or [THEN]
respectively. While doing so, [IFDEF] does not get compiled to [IF],
so we need to handle nested [IFDEF] to account for its [ELSE] or [THEN].

This fixes [IFDEF] not disabling the full section of code.

Thanks to Segher for pointing me in the right direction.

Signed-off-by: Andreas Färber <andreas.faerber@web.de>
Signed-off-by: Segher Boessenkool <segher@kernel.crashing.org>


git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@979 f158a5a8-5612-0410-a976-696ce0be7e32