forth: Fix MMU available pretty-printing for sparc64
On sparc64 a virtual address or size value is contained within one stack cell
but that makes two property cells. The columns for the /virtual-memory
"available" property thus looked wrong.

Do read the number of cells from #address-cells but make sure the size is
at least one to not break ppc.

Cc: Tarl Neustaedter <tarl-b2@tarl.net>
Signed-off-by: Andreas Färber <andreas.faerber@web.de>


git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@984 f158a5a8-5612-0410-a976-696ce0be7e32