Implement ofmem_map_io() within OFMEM as a function to map I/O memory (similar to SPARC32's private map_io() function).
As a consequence we need to maintain a separate I/O range within OFMEM, and add management functions similar to those that
exist for the virtual and physical ranges. At the same time, we introduce a couple of new OFMEM arch functions called
ofmem_arch_get_iomem_base() and ofmem_arch_get_iomem_top() to control the range of allocatable I/O memory which is again
similar to the existing method used for the virtual and physical ranges.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@siriusit.co.uk>
Tested-by: Blue Swirl <blauwirbel@gmail.com>
Reviewed-by: Blue Swirl <blauwirbel@gmail.com>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1007 f158a5a8-5612-0410-a976-696ce0be7e32