Implement new v3_memalloc romvec interface for allocating aligned memory.
When attempting to load kadb from a Solaris 8 CD at boot, it seems that in order
for kadb to allocate enough memory to load the kernel (and its modules) the
/openprom/aligned-allocator property must be present with length zero.

This has the additional side-effect that OpenBIOS memory allocations are
redirected to a new romvec offset when this property is detected, which appears
to have a similar interface to obp_dumb_memalloc() but with an extra
alignment parameter. Hence we include a simple implementation which seems
to be enough to be able to get kadb to load and attach to the kernel.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@siriusit.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1035 f158a5a8-5612-0410-a976-696ce0be7e32