Make select-dev and unselect-dev compatible with OBP.
From the 1275 errata for section H.8, select-dev and unselect-dev are
pre Open Firmware, but not the same as open-dev/device-end.  Make
them Sun OBP compatible by pulling the functionality for select-dev/
unselect-dev out of (and simplifying) begin-package/end-package.

For the Sun OBP definition of begin-package/end-package, along with
descriptions and example usage, see the "Debugging and Testing FCode
Programs" chapter of "Writing FCode 3.x Programs".

Signed-off-by: Bob Breuer <breuerr@mc.net>
Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@siriusit.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1046 f158a5a8-5612-0410-a976-696ce0be7e32