Improve dir word by reducing complexity and adding some more diagnostics.
Since the OF path resolution algorithm automatically passes anything following
a ':' as an argument to the specified device, we don't need to bother with
handling this ourselves. As a side effect, the auxiliary word split-path-device
is no longer needed as we can isolate the path by doing a simple split on ','.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1071 f158a5a8-5612-0410-a976-696ce0be7e32