Rework mac-parts.c to use CHRP-compliant partition search, followed by Apple OF partition search.
The confusion from the previous refactoring of mac-parts.c came from the fact
that it must support searching for a boot partition using the algorithm
specified in the OF CHRP bindings supplement, as well as Apple's brute-force
partition-type search.

This patch corrects mac-parts.c so that it will first attempt to find a CHRP
boot partition, and if it does not exist or fails to boot, fall back to the
first partition that would be considered valid by Apple's OF if one exists.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1076 f158a5a8-5612-0410-a976-696ce0be7e32