mac-parts.c: Fix /packages/mac-parts boot when the package arguments are empty.
Commit r1072 removed a couple of NULL pointer checks for package arguments,
however if the arguments are empty in Forth then they are converted back to a
NULL pointer which causes an exception when dereferenced.

Add the checks back in which fixes boot of the debian_lenny_powerpc_small.qcow2
test image with quik once again. Based upon an earlier patch by Alexander Graf
<agraf@suse.de>.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>
CC: Alexander Graf <agraf@suse.de>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1100 f158a5a8-5612-0410-a976-696ce0be7e32