SPARC32: Move Forth dictionary location to a fixed location.
Instead of using the automatic allocator to determine the location of the
Forth dictionary, allocate it at a fixed location just underneath the ROM
itself at 0xffd00000.

This patch also prevents corruption within the RAM of the Forth machine
by ensuring that we allocate virtual memory outside of OpenBIOS itself,
plus reduces the size of the Forth dictionary to 128K so that it meets
the memory constraints required to boot Linux and also kadb from older
versions of SunOS (as reported by Artyom Tarasenko).

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1114 f158a5a8-5612-0410-a976-696ce0be7e32