Remove /chosen "memory" property from the default device tree.
Similar to the earlier commit for "mmu", do the same for /chosen "memory"
property. Hence all architectures that want the "pretty" memory properties
can provide suitable ihandles if required, and those that don't will not
fail with a NULL pointer dereference from a zero ihandle.

While we're here, correct the SPARC32 initialiser to point to the correct
/virtual-memory node for "mmu" so that once again we can get "pretty"
memory properties on SPARC32.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1120 f158a5a8-5612-0410-a976-696ce0be7e32