disk-label.c: Fix opening of raw block devices
If open-dev is called on a raw block device such as cdrom, the current code
will always attempt to interpose a partition handler for the first valid
partition table type it finds.

This is incorrect when opening a raw block device directly as it means we
"fall into" the first valid partition that can be found, rather than using
the full raw device. Fix this by only attempting to interpose a partition
handler if the package arguments are not empty.

This fixes the NetBSD SPARC64 bootloader as reported by
Martin Husemann in https://bugs.launchpad.net/qemu/+bug/1169856.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1127 f158a5a8-5612-0410-a976-696ce0be7e32