video: rename qemu-video-* words to openbios-video-*
Now that the second half of the display rework is complete, the qemu-video-*
words are no longer QEMU-specific. This is because we now initialise the
defaults to the value of the VGA_DEFAULT_* constants and only attempt to
override them with the values supplied by the QEMU firmware interface if
built with CONFIG_QEMU enabled.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1195 f158a5a8-5612-0410-a976-696ce0be7e32