pci/vga: move PCI framebuffer map functions info Forth
Here we move the code to map the PCI framebuffer into Forth. This is done by
introducing a "pci-map-in" word on the /pci node to enable Forth code to
perform memory-mapping as required.

This is a slight departure from the specification which requires a "map-in"
word, but this is because the OpenBIOS PCI bus enumerator *always* configures
BARs regardless of whether or not an FCode payload is detected. In order to
avoid having to rewrite the bus enumeration code in Forth, we instead create
a "pci-bar>pci-region" word which takes the BAR register as a parameter and
returns the configured BAR bus address and size. This can then be passed to
the "pci-map-in" word in order to do the work.

Finally with this in place, we can remove the mapping that takes place in
vga_vbe_init().

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1207 f158a5a8-5612-0410-a976-696ce0be7e32