tcx/cgthree: remove hard OpenBIOS dependency from driver
Since QEMU 1.7, the OpenBIOS TCX framebuffer FCode ROM is made available to
Open Firmware upon boot. However, this causes a boot failure for people who
use a Sun OBP firmware with QEMU since the FCode throws an exception trying
to access some OpenBIOS-specific variables. This causes an invalid device
node to be created in the tree, causing a crash on boot as reported by Peter
Bartoli and Artyom Tarasenko.

Fix both drivers so that if OpenBIOS is not detected, we fall back to the QEMU
framebuffer defaults of 1024x768x8 which should generate a valid device node
and avoid the crash.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1246 f158a5a8-5612-0410-a976-696ce0be7e32