OFMEM: allocate retained memory top downwards
Not only does this match the behaviour of the other routines, but it also
fixes a bug whereby the first free physical memory location was 0x0 and
so page zero was being returned for use by SUNW,retain clients. Due to the
way in which PCI/ISA devices are mapped in QEMU (e.g. IDE ioports) then
clients could inadvertantly start writing to device registers.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1255 f158a5a8-5612-0410-a976-696ce0be7e32