OFMEM: improve remove_range_() function algorithm
Rework the algorithm used in remove_range_() in order to make it similar to
the algorithm used in unmap_pages(). This has several benefits: in particular
the unmap_pages()/remove_range_() functions now have the same behaviour, the
code is noticeably simpler, and it also correctly handles the case where the
start address of the range to be removed doesn't match the range exactly (such
as occurs when attempting to boot NextStep under SPARC32).

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1281 f158a5a8-5612-0410-a976-696ce0be7e32