SPARC64: fix up translations property mode to make it a valid TTE
NetBSD attempts to parse the translations property in order to set up the
virtual to physical translations for the kernel. Alter the mode cell to
include the physical address and valid bit so that it represents a real
TTE entry for the given start adddress.

This is confirmed by checking real dumps from prtconf examples, plus the
NetBSD kernel now starts to boot.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1296 f158a5a8-5612-0410-a976-696ce0be7e32