SPARC64: only map 8MB RAM on startup
Due to a (bug?) in OpenBSD's pmap_bootstrap(), the kernel attempts to locate
its text segment and size by searching the translations property for continuous
regions in use starting from the kernel load address.

Unfortunately if we have already mapped memory above the kernel load address of
0x1000000 then the kernel extends the text region way beyond the text segement
causing an exception when eventually trying to set non-existent mappings in the
TLB.

Fix this by only mapping 8MB of RAM so that we guarantee that the region above
0x1000000 will always be free, and hence the text/data regions will be discrete
so that the sizing algorithms return the correct answer.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1298 f158a5a8-5612-0410-a976-696ce0be7e32