package.fs: fix "child" word being called with a NULL phandle
Solaris 9 relies on an undocumented feature that calling "child" with
a NULL phandle is equivalent to using the phandle of the device tree
root node.

Make sure that we emulate the same behaviour to avoid a crash when
dereferencing a NULL pointer.

This patch fixes Solaris 9 boot on SPARC32 and is based upon a
proof-of-concept patch submitted by Artyom Tarasenko <atar4qemu@gmail.com>.

Signed-off-by: Mark Cave-Ayland <mark.cave-ayland@ilande.co.uk>

git-svn-id: svn://openbios.org/openbios/trunk/openbios-devel@1328 f158a5a8-5612-0410-a976-696ce0be7e32